lifetime: circular dependency; Fix #77
