+++ lib/rucio/common/policy.py
@@ -16,20 +16,13 @@ import os
 import logging
 import sys
 
+from functools import wraps
+
 from ConfigParser import NoOptionError, NoSectionError
 from dogpile.cache import make_region
 from dogpile.cache.api import NoValue
 
-import rucio.core.did
-import rucio.core.lock
-import rucio.core.rule
-
-from rucio.db.sqla.constants import DIDType
-from rucio.db.sqla.session import read_session, transactional_session
 from rucio.common.config import config_get
-from rucio.common.exception import DataIdentifierNotFound
-from rucio.core.account import has_account_attribute
-from rucio.core.rse import list_rse_attributes
 
 REGION = make_region().configure('dogpile.cache.memory',
                                  expiration_time=1800)
@@ -79,92 +72,17 @@ def get_lifetime_policy():
     return lifetime_dict
 
 
-@read_session
-def get_scratch_policy(account, rses, lifetime, session=None):
-    """
-    ATLAS policy for rules on SCRATCHDISK
-
-    :param account:  Account of the rule.
-    :param rses:     List of RSEs.
-    :param lifetime: Lifetime.
-    :param session:  The database session in use.
-    """
-
+def policy_filter(function):
+    mapping = {'atlas': ['get_scratch_policy', 'archive_localgroupdisk_datasets']}
     vo_name = get_vo()
-    scratchdisk_lifetime = get_scratchdisk_lifetime()
-    if vo_name == 'atlas':
-        # Check SCRATCHDISK Policy
-        if not has_account_attribute(account=account, key='admin', session=session) and (lifetime is None or lifetime > 60 * 60 * 24 * scratchdisk_lifetime):
-            # Check if one of the rses is a SCRATCHDISK:
-            if [rse for rse in rses if list_rse_attributes(rse=None, rse_id=rse['id'], session=session).get('type') == 'SCRATCHDISK']:
-                lifetime = 60 * 60 * 24 * scratchdisk_lifetime - 1
-    return lifetime
+    if vo_name in mapping and function.__name__ in mapping[vo_name]:
+        @wraps(function)
+        def new_funct(*args, **kwargs):
+            return function(*args, **kwargs)
+        print function.__name__
+        return new_funct
 
-
-@transactional_session
-def archive_localgroupdisk_datasets(scope, name, session=None):
-    """
-    ATLAS policy to archive a dataset which has a replica on LOCALGROUPDISK
-
-    :param scope:    Scope of the dataset.
-    :param name:     Name of the dataset.
-    :param session:  The database session in use.
-    """
-
-    vo_name = get_vo()
-    if vo_name != 'atlas':
+    @wraps(function)
+    def null_funct(*args, **kwargs):
         return
-
-    rses_to_rebalance = []
-
-    # Check if the archival dataset already exists
-    try:
-        rucio.core.did.get_did(scope='archive', name=name, session=session)
-        return
-    except DataIdentifierNotFound:
-        pass
-
-    # Check if the dataset has a rule on a LOCALGROUPDISK
-    for lock in rucio.core.lock.get_dataset_locks(scope=scope, name=name, session=session):
-        if 'LOCALGROUPDISK' in lock['rse']:
-            rses_to_rebalance.append({'rse': lock['rse'], 'account': lock['account']})
-    # Remove duplicates from list
-    rses_to_rebalance = [dict(t) for t in set([tuple(sorted(d.items())) for d in rses_to_rebalance])]
-
-    # There is at least one rule on LOCALGROUPDISK
-    if rses_to_rebalance:
-        content = [x for x in rucio.core.did.list_content(scope=scope, name=name, session=session)]
-        if len(content) > 0:
-            # Create the archival dataset
-            did = rucio.core.did.get_did(scope=scope, name=name, session=session)
-            meta = rucio.core.did.get_metadata(scope=scope, name=name, session=session)
-            new_meta = {k: v for k, v in meta.items() if k in ['project', 'datatype', 'run_number', 'stream_name', 'prod_step', 'version', 'campaign', 'task_id', 'panda_id'] and v is not None}
-            rucio.core.did.add_did(scope='archive',
-                                   name=name,
-                                   type=DIDType.DATASET,
-                                   account=did['account'],
-                                   statuses={},
-                                   meta=new_meta,
-                                   rules=[],
-                                   lifetime=None,
-                                   dids=[],
-                                   rse=None,
-                                   session=session)
-            rucio.core.did.attach_dids(scope='archive', name=name, dids=content, account=did['account'], session=session)
-            if not did['open']:
-                rucio.core.did.set_status(scope='archive', name=name, open=False, session=session)
-
-            for rse in rses_to_rebalance:
-                rucio.core.rule.add_rule(dids=[{'scope': 'archive', 'name': name}],
-                                         account=rse['account'],
-                                         copies=1,
-                                         rse_expression=rse['rse'],
-                                         grouping='DATASET',
-                                         weight=None,
-                                         lifetime=None,
-                                         locked=False,
-                                         subscription_id=None,
-                                         ignore_account_limit=True,
-                                         ignore_availability=True,
-                                         session=session)
-            logging.debug('Re-Scoped %s:%s' % (scope, name))
+    return null_funct
